+++ R/ictr_data.R
@@ -1,0 +1,16 @@
+#' TODO: Document
+#'
+#' @author Kalash Singhal.
+#' @source TODO
+#' @family ICTR datasets
+#' @examples
+#' ictr_inputs
+"ictr_inputs"
+#' TODO: Document
+#'
+#' @author Kalash Singhal.
+#' @source TODO
+#' @family ICTR datasets
+#' @examples
+#' ictr_companies
+"ictr_companies"