+++ R/pctr_data.R
@@ -1,16 +1,17 @@
 #' TODO add title
 #'
-#' @family PCTR datasets
-#'
+#' @author Tilman Trompke.
 #' @source TODO add source
+#' @family PCTR datasets
 #' @examples
 #' pctr_ecoinvent_co2
 "pctr_ecoinvent_co2"
 
 #' TODO add title
 #'
-#' @family PCTR datasets
+#' @author Tilman Trompke.
 #' @source TODO add source
+#' @family PCTR datasets
 #' @examples
 #' pctr_companies
 "pctr_companies"