+++ R/pctr_score_companies.R
@@ -1,8 +1,13 @@
-#' Title TODO
+#' Aggregate the products' scores for each company
 #'
-#' Description TODO
+#' @description
+#' Calculates on a company-level the percentage of products that are in low /
+#' medium / high transition risk due to the products' relative carbon footprint.
 #'
-#' Details (optional) TODO
+#' Activities are mapped with the companies' products using common columns
+#' present in dataframes `scored_activities` and `companies`
+#'
+#' @author Tilman Trompke.
 #'
 #' @param scored_activities A [data.frame]. The output of
 #'   [pctr_score_activities].
@@ -10,13 +15,25 @@
 #'
 #' @family PCTR functions
 #'
-#' @return A [data.frame] with columns ... TODO
+#' @return A [data.frame] with three rows for each company and these columns:
+#'   * `company_id`
+#'   * `score`
+#'   * `share_all`
+#'   * `share_unit`
+#'   * `share_unit_sec`
+#'   where `share_all`, `share_unit`, and `share_unit_sec` holds the
+#'   aggregated scores in percentage.
 #' @export
 #'
 #' @examples
+#' library(dplyr, warn.conflicts = FALSE)
+#'
+#' one_company <- pctr_companies |> filter(company_id %in% first(company_id))
+#' one_company
+#'
 #' pctr_ecoinvent_co2 |>
 #'   pctr_score_activities(low_threshold = 0.3, high_threshold = 0.7) |>
-#'   pctr_score_companies(pctr_companies)
+#'   pctr_score_companies(one_company)
 pctr_score_companies <- function(scored_activities, companies) {
   # Intermediate steps:
   # * Combine company-level information with LCA info from ecoinvent.
@@ -24,7 +41,7 @@ pctr_score_companies <- function(scored_activities, companies) {
 
   # join by activity_product_uuid and other joint columns from companies with scored_activities
   companies_scores <- companies |>
-    inner_join(scored_activities, by = c("activity_product_uuid", "ei_activity", "unit"))
+    left_join(scored_activities, by = c("activity_product_uuid", "ei_activity", "unit"))
 
   # scores in comparison to all products
   scores_all <- companies_scores |>
@@ -53,7 +70,7 @@ pctr_score_companies <- function(scored_activities, companies) {
   # create dataset sceleton
   dt_sceleton <- tibble(
     company_id = rep(unique(companies_scores$company_id), each = 3),
-    score = rep(c("high", "medium", "low"), 3),
+    score = rep(c("high", "medium", "low"), length(unique(companies_scores$company_id))),
   )
 
   # join scores with dt_sceleton so that each company is shown with 3 rows for low, medium, and high, even if the share is 0.