+++ R/utils.R
@@ -35,26 +35,27 @@ mvp_paths <- function(pattern = NULL) {
 #' use_refactoring_checklist()
 use_refactoring_checklist <- function() {
   glue::glue("
-  - [ ] Confirm the data is public.
-  - [ ] Add mvp .Rmd.
+  - [ ] Add the mvp .Rmd.
   - [ ] Redirect `here` to access inputs.
-  - [ ] Render and assess viability. If viable contine else ask for changes.
-  - [ ] `use_article()`. with mvp .Rmd as child document.
+  - [ ] Render and assess viability. If not viable ask for changes.
+  - [ ] `use_article()`. Reuse the mvp .Rmd as a `child` document.
   - [ ] Document authorship in DESCRIPTION, mvp, and article.
   - [ ] Snapshot important objects in the rendering environment or the mvp.
-  - [ ] In data-raw/ `use_data()`.
-  - [ ] `use_r('*_data')`.
-  - [ ] Reuse data in .Rmd files.
+  - [ ] `use_data()`.
+  - [ ] Document data minimally to avoid R CMD issues and to show an example.
+  - [ ] Reuse data in the mvp .Rmd.
   - [ ] Polish and publish article.
-  - [ ] Bump version.
   - [ ] Announce new Article on Slack.
-  - [ ] Discuss the API with the author.
   - [ ] Refactor: Extract functions in place.
-  - [ ] Refactor: Move functions to R/ and address R CMD check issues.
-  - [ ] Reorganize documentation (e.g. use man/fragments) and flag TODO.
-  - [ ] Update CODEOWNWERS.
-  - [ ] Ask authors to help with documentation.
+  - [ ] Refactor: Move functions to R/ and document @author.
+  - [ ] Refactor: Address R CMD check issues.
+  - [ ] Ask authors to document data and functions, then prune the article.
   - [ ] Snapshot each feature and remove the general mvp-level snapshot.
-  - [ ] Refactor: Move mvp to article.
+  - [ ] Move mvp to article.
+  - [ ] Add basic characterization tests and discuss with the author.
+  - [ ] Discuss the API with the author.
+  - [ ] DRY code across mvps.
+  - [ ] DRY documentation (e.g. via man/fragments).
+  - [ ] Update CODEOWNWERS.
   ")
 }